back/requirements.txt
django==5.1
djangorestframework==3.15.2
channels==4.1.0
redis==5.0.8
channels[daphne]==4.1.0